requirements.txt
mdpdf
requests
pycryptodome
pypiwin32
pyinstaller